isl: bump to version 0.18
Now that we no longer support Graphite optimizations in gcc 4.9 (which
required ISL 0.14.1), we can bump ISL to a newer version, which allows
to support Graphite optimizations in GCC >= 5.x.

The comment about the version being limited to 0.14.1 because of cloog
is no longer relevant, so it is removed.

Our libtool patch now applies properly, so we can remove
ISL_LIBTOOL_PATCH = NO.

Signed-off-by: Andrey Yurovsky <yurovsky@gmail.com>
[Thomas:
 - remove ISL_LIBTOOL_PATCH = NO
 - improved commit log]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>