bcusdk: enable tpuarts support in eibd
This commit will enable tpuarts in eibd as well. tpuarts is as
important as FT1.2 support, which is already enabled by default.

This causes a 8 KB increase of the installed size in /usr/bin (728 KB
instead of 720 KB for /usr/bin), so not enough to justify a separate
configuration option.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>