pdbg: bump version to latest
I got this error when attempting to build:

 aclocal: error: couldn't open directory 'm4': No such file or directory

So we now have the post patch hook to ensure the m4 directory is
present.

In addition, the package now sets a variable with the Git SHA1 so that
the binary contains the expected output for --version.

Specific configurations of uClibc cause a build failure in pdbg. This commit
adds a patch to remove the definition of the offending macro (it is unused in
the source).

 In file included from usr/include/stdio.h:71:0,
                  from libpdbg/fake.c:17:
 include/bits/uClibc_stdio.h:149:16: error: expected identifier or ‘(’ before ‘;’ token
   void *__unused;    /* Placeholder for codeset binding. */

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>